APP-1921 Cache probe responses
Be a bit more 'daemon like' and remember previous probe responses.
This helps when the network is a bit loaded and we might miss
later responses; i.e. particularly in the situation where we
load the page, start a bunch of streams, then refresh, the cameras
should no longer disappear.

However, it shouldn't be used in morse_feed until we have
better frontend logic for reporting unresponsive cameras
(cf the current 'ONVIF query failure').

Approved-by: Arien Judge
Approved-by: Milad Mohtashamirad